llm_projects/data_analyis/requirements.txt
streamlit>=1.28.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain>=0.3.0
langgraph>=0.0.20
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0
openai>=1.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu
